requirements.txt
# OpenInterpretability - Advanced LLM Behavior Analysis and Interpretability Platform
# Author: Nik Jois <nikjois@llamasearch.ai>

# Core dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.4.0
pydantic-settings>=2.0.0

# AI/ML dependencies
openai>=1.3.0
anthropic>=0.7.0
numpy>=1.24.0

# Async and HTTP
aiohttp>=3.9.0
asyncio-throttle>=1.0.2

# Data handling
pyyaml>=6.0
python-multipart>=0.0.6

# Security and authentication
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0

# Caching and storage
redis>=5.0.0

# Database and data tools
sqlite-utils>=3.34.0
datasette>=0.64.3
databases[aiosqlite]>=0.8.0
alembic>=1.12.0

# CLI and command line tools
click>=8.1.0
rich>=13.6.0
typer>=0.9.0

# Configuration management
python-dotenv>=1.0.0

# Logging and monitoring
structlog>=23.2.0

# Development and testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
httpx>=0.25.0

# Code quality
black>=23.10.0
isort>=5.12.0
flake8>=6.1.0
mypy>=1.6.0

# Documentation
sphinx>=7.2.0
sphinx-rtd-theme>=1.3.0

# Optional dashboard dependencies
streamlit>=1.28.0
plotly>=5.17.0
pandas>=2.1.0

# Additional utilities
python-dateutil>=2.8.0
tqdm>=4.66.0
tabulate>=0.9.0 